chore: move Icon.svelte from icons/ to components/
