+++ packages/mochi/src/components/Icon.svelte
