+++ packages/mochi/src/icons/chevron-right.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-  import Icon from './Icon.svelte';
+  import Icon from '../components/Icon.svelte';
   let props = $props();
   const iconNode = [['path', { d: 'm9 18 6-6-6-6' }]];
 </script>
