+++ packages/mochi/src/icons/lock.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-  import Icon from './Icon.svelte';
+  import Icon from '../components/Icon.svelte';
   let props = $props();
   const iconNode = [
     ['rect', { width: '18', height: '11', x: '3', y: '11', rx: '2', ry: '2' }],
